refactor: simplify module boundaries
