+++ baize-flux-api/src/main/java/com/baize/flux/api/table/catalog/exception/CatalogException.java
@@ -1,7 +1,7 @@
 package com.baize.flux.api.table.catalog.exception;
 
-import com.baize.flux.common.exception.FluxRuntimeException;
-import com.baize.flux.common.exception.error.FluxAPIErrorCode;
+import com.baize.flux.api.exception.FluxRuntimeException;
+import com.baize.flux.api.exception.error.FluxApiErrorCode;
 
 /**
  * Catalog 操作异常。
@@ -11,17 +11,17 @@ public class CatalogException extends FluxRuntimeException {
     private static final long serialVersionUID = 1L;
 
     public CatalogException(String message) {
-        super(FluxAPIErrorCode.CATALOG_INITIALIZE_FAILED, message);
+        super(FluxApiErrorCode.CATALOG_INITIALIZE_FAILED, message);
     }
 
     public CatalogException(
             String message,
             Throwable cause) {
 
-        super(FluxAPIErrorCode.CATALOG_INITIALIZE_FAILED, message, cause);
+        super(FluxApiErrorCode.CATALOG_INITIALIZE_FAILED, message, cause);
     }
 
     public CatalogException(Throwable cause) {
-        super(FluxAPIErrorCode.CATALOG_INITIALIZE_FAILED, cause);
+        super(FluxApiErrorCode.CATALOG_INITIALIZE_FAILED, cause);
     }
 }
