--- baize-flux-connectors/baize-flux-connector-common/pom.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xmlns="http://maven.apache.org/POM/4.0.0"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>com.baize.flux</groupId>
-        <artifactId>baize-flux-connectors</artifactId>
-        <version>1.0.0</version>
-    </parent>
-
-    <artifactId>baize-flux-connector-common</artifactId>
-    <version>1.0.0</version>
-
-
-</project>
