--- baize-flux-connectors/baize-flux-connector-common/src/main/java/com/baize/flux/connectors/common/Test.java
@@ -1,4 +1,0 @@
-package com.baize.flux.connectors.common;
-
-public class Test {
-}
