+++ baize-flux-connectors/pom.xml
@@ -15,43 +15,7 @@
     <packaging>pom</packaging>
 
     <modules>
-        <module>baize-flux-connector-common</module>
         <module>baize-flux-connector-jdbc</module>
     </modules>
 
-    <dependencyManagement>
-        <dependencies>
-            <dependency>
-                <groupId>com.baize.flux</groupId>
-                <artifactId>baize-flux-api</artifactId>
-                <version>${project.version}</version>
-            </dependency>
-        </dependencies>
-    </dependencyManagement>
-
-
-    <dependencies>
-        <dependency>
-            <groupId>com.baize.flux</groupId>
-            <artifactId>baize-flux-api</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>com.baize.flux</groupId>
-            <artifactId>baize-flux-framework</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>com.h2database</groupId>
-            <artifactId>h2</artifactId>
-            <version>2.2.224</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>junit</groupId>
-            <artifactId>junit</artifactId>
-            <version>${junit.version}</version>
-            <scope>test</scope>
-        </dependency>
-    </dependencies>
 </project>
