+++ baize-flux-framework/src/main/java/com/baize/flux/framework/connector/ConnectorException.java
@@ -1,7 +1,7 @@
 package com.baize.flux.framework.connector;
 
-import com.baize.flux.common.exception.FluxRuntimeException;
-import com.baize.flux.common.exception.error.FluxAPIErrorCode;
+import com.baize.flux.api.exception.FluxRuntimeException;
+import com.baize.flux.api.exception.error.FluxApiErrorCode;
 
 /**
  * Connector 发现、校验或准备失败。
@@ -12,13 +12,13 @@ public class ConnectorException
     private static final long serialVersionUID = 1L;
 
     public ConnectorException(String message) {
-        super(FluxAPIErrorCode.CONNECTOR_INITIALIZE_FAILED, message);
+        super(FluxApiErrorCode.CONNECTOR_INITIALIZE_FAILED, message);
     }
 
     public ConnectorException(
             String message,
             Throwable cause) {
 
-        super(FluxAPIErrorCode.CONNECTOR_INITIALIZE_FAILED, message, cause);
+        super(FluxApiErrorCode.CONNECTOR_INITIALIZE_FAILED, message, cause);
     }
 }
