+++ baize-flux-framework/src/main/java/com/baize/flux/framework/factory/FactoryException.java
@@ -1,7 +1,7 @@
 package com.baize.flux.framework.factory;
 
-import com.baize.flux.common.exception.FluxRuntimeException;
-import com.baize.flux.common.exception.error.FluxAPIErrorCode;
+import com.baize.flux.api.exception.FluxRuntimeException;
+import com.baize.flux.api.exception.error.FluxApiErrorCode;
 
 /**
  * Factory 发现、校验或创建失败。
@@ -11,12 +11,12 @@ public class FactoryException extends FluxRuntimeException {
     private static final long serialVersionUID = 1L;
 
     public FactoryException(String message) {
-        super(FluxAPIErrorCode.FACTORY_INITIALIZE_FAILED, message);
+        super(FluxApiErrorCode.FACTORY_INITIALIZE_FAILED, message);
     }
 
     public FactoryException(
             String message,
             Throwable cause) {
-        super(FluxAPIErrorCode.FACTORY_INITIALIZE_FAILED, message, cause);
+        super(FluxApiErrorCode.FACTORY_INITIALIZE_FAILED, message, cause);
     }
 }
